+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - **Breaking**: Replace `FragmentStoreBuilder` with `FeaturesBuilder`
 - **Breaking**: Rename `FragmentFlowStore` to `FragmentStore`, `FragmentState` to `FragmentOutput`, `FragmentFlowState` to `FragmentState`
 - **Breaking**: Making `FeatureFactory` into an abstract class and modifying the `initialize` API to use `Params` wrapper type
+- **Breaking**: Move factory functions from `FlowAction` to `CoroutineAction`
 
 ## [0.7.1] - June 28, 2022
 - **Breaking**: Rename `FragmentBindingBuilder` to `FragmentStoreBuilder`